Find by by your place

+++ src/components/UtilityTableRow.js
@@ -1,8 +1,9 @@
-import React, {Component} from 'react';
-import {Link} from 'react-router-dom';
+import React, { Component } from 'react';
+import { Link } from 'react-router-dom';
+import Card from 'react-bootstrap/Card';
 import axios from 'axios';
 import Button from 'react-bootstrap/Button';
-import {connect} from "react-redux";
+import { connect } from "react-redux";
 
 class StudentTableRow extends Component {
 
@@ -12,12 +13,18 @@ class StudentTableRow extends Component {
     }
 
     deleteStudent() {
-        axios.delete('http://localhost:4000/utility/delete-utility/' + this.props.obj._id)
-            .then((res) => {
-                alert('successfully deleted');
-            }).catch((error) => {
-            console.log(error)
-        })
+        var r = window.confirm("Are you sure?");
+        if (r == true) {
+            axios.delete('http://localhost:4000/utility/delete-utility/' + this.props.obj._id)
+                .then((res) => {
+                    alert('successfully deleted');
+                    window.location.reload();
+                }).catch((error) => {
+                    console.log(error)
+                })
+        } else {
+            console.log('Item is not deleted');
+        }
     }
 
     render() {
@@ -26,22 +33,30 @@ class StudentTableRow extends Component {
         const logged = this.props.auth.user.id;
         if (isAdmin === true || logged === this.props.obj.creatorId) {
             return (
-                <tr>
-                    <td>{this.props.obj.title}</td>
-                    <td>{this.props.obj.category}</td>
-                    <td>{this.props.obj.description}</td>
-                    <td>{this.props.obj.price}</td>
-                    <td>{this.props.obj.phone}</td>
-                    <td>{this.props.obj.town}</td>
-                    <td>{this.props.obj.townLng}</td>
-                    <td>{this.props.obj.townLat}</td>
-                    <td>
-                        <Link className="edit-link" to={"/edit-utility/" + this.props.obj._id}>
-                            Edit
-                        </Link>
-                        <Button onClick={this.deleteStudent} size="sm" variant="danger">Delete</Button>
-                    </td>
-                </tr>
+                // <tr>
+                <Card>
+  <Card.Header as="h5">{this.props.obj.title}</Card.Header>
+  <Card.Body>
+    <Card.Title>{this.props.obj.category}</Card.Title>
+    <Card.Text>
+      With supporting text below as a natural lead-in to additional content.
+    </Card.Text>
+    <Button variant="primary">Go somewhere</Button>
+  </Card.Body>
+</Card>
+                    // <td>{this.props.obj.title}</td>
+                    // <td>{this.props.obj.category}</td>
+                    // <td>{this.props.obj.description}</td>
+                    // <td>{this.props.obj.price}</td>
+                    // <td>{this.props.obj.phone}</td>
+                    // <td>{this.props.obj.town}</td>
+                    // <td>
+                        // <Link className="edit-link" to={"/edit-utility/" + this.props.obj._id}>
+                        //     Edit
+                        // </Link>
+                        // <Button onClick={this.deleteStudent} size="sm" variant="danger">Delete</Button>
+                    // {/* </td> */}
+                // {/* </tr> */}
             );
         } else {
             return (
@@ -52,8 +67,6 @@ class StudentTableRow extends Component {
                     <td>{this.props.obj.price}</td>
                     <td>{this.props.obj.phone}</td>
                     <td>{this.props.obj.town}</td>
-                    <td>{this.props.obj.townLng}</td>
-                    <td>{this.props.obj.townLat}</td>
                     {/*<td>*/}
                     {/*    <Link className="edit-link" to={"/edit-utility/" + this.props.obj._id}>*/}
                     {/*        Edit*/}
